apps/web-backend/pyproject.toml
[project]
name = "web-backend"
version = "0.2.0"
description = "FastAPI web backend for serving ML models"
requires-python = ">=3.12"
dependencies = [
    "ml-core",
    "fastapi>=0.110.0",
    "httpx>=0.25.0",
    "pydantic>=2.12.5",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.0.0",
    "uvicorn>=0.41.0",
    "python-dotenv>=1.2.1",
    "psycopg2-binary>=2.9.11",
    # Auth dependencies
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
]

[dependency-groups]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.25.0",
]

[tool.uv.sources]
ml-core = { workspace = true }

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"training".msg = "Don't import across app boundaries"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"